+++ src/common/services/logger.service.ts
@@ -1,43 +1,56 @@
+import * as clc from 'cli-color';
 import { NestEnvironment } from '../enums/nest-environment.enum';
 
 declare const process;
-import * as clc from 'cli-color';
 
 export class Logger {
-    private static mode = NestEnvironment.RUN;
+    private static lastTimestamp = null;
+    private static contextEnv = NestEnvironment.RUN;
     private readonly yellow = clc.xterm(3);
 
-    constructor(private context: string) {}
+    constructor(
+        private readonly context: string,
+        private readonly printTimestamps = false) {}
 
     public static setMode(mode: NestEnvironment) {
-        this.mode = mode;
+        this.contextEnv = mode;
     }
 
     public log(message: string) {
-        this.logMessage(message, clc.green);
+        this.printMessage(message, clc.green);
     }
 
     public error(message: string, trace = '') {
-        this.logMessage(message, clc.red);
+        this.printMessage(message, clc.red);
         this.printStackTrace(trace);
     }
 
     public warn(message: string) {
-        this.logMessage(message, clc.yellow);
+        this.printMessage(message, clc.yellow);
     }
 
-    private logMessage(message: string, color: (msg: string) => string) {
-        if (Logger.mode === NestEnvironment.TEST) return;
+    private printMessage(message: string, color: (msg: string) => string) {
+        if (Logger.contextEnv === NestEnvironment.TEST) return;
 
         process.stdout.write(color(`[Nest] ${process.pid}   - `));
         process.stdout.write(`${new Date(Date.now()).toLocaleString()}   `);
         process.stdout.write(this.yellow(`[${this.context}] `));
         process.stdout.write(color(message));
+
+        this.printTimestamp();
         process.stdout.write(`\n`);
     }
 
+    private printTimestamp() {
+        const includeTimestamp = Logger.lastTimestamp && this.printTimestamps;
+        if (includeTimestamp) {
+            process.stdout.write(this.yellow(` +${Date.now() - Logger.lastTimestamp}ms`));
+        }
+        Logger.lastTimestamp = Date.now();
+    }
+
     private printStackTrace(trace: string) {
-        if (Logger.mode === NestEnvironment.TEST || !trace) return;
+        if (Logger.contextEnv === NestEnvironment.TEST || !trace) return;
 
         process.stdout.write(trace);
         process.stdout.write(`\n`);